+++ src/_packages/utils.index.ts
@@ -1,7 +1,6 @@
 // ⚠️ WARNING: This code has been generated so that any manual changes will be overwritten
 // `@promptbook/utils`
 
-import { PROMPTBOOK_VERSION } from '../version';
 import { renderPromptbookMermaid } from '../conversion/prettify/renderPipelineMermaidOptions';
 import { extractParameterNamesFromPromptTemplate } from '../conversion/utils/extractParameterNamesFromPromptTemplate';
 import { extractVariables } from '../conversion/utils/extractVariables';
@@ -11,9 +10,6 @@ import { forEachAsync } from '../execution/utils/forEachAsync';
 import { isValidJsonString } from '../formats/json/utils/isValidJsonString';
 import { extractBlock } from '../postprocessing/utils/extractBlock';
 import { $currentDate } from '../utils/$currentDate';
-import { clonePipeline } from '../utils/clonePipeline';
-import { deepClone } from '../utils/deepClone';
-import { $deepFreeze } from '../utils/deepFreeze';
 import { $getGlobalScope } from '../utils/environment/$getGlobalScope';
 import { $isRunningInBrowser } from '../utils/environment/$isRunningInBrowser';
 import { $isRunningInNode } from '../utils/environment/$isRunningInNode';
@@ -22,16 +18,14 @@ import { countCharacters } from '../utils/expectation-counters/countCharacters';
 import { countLines } from '../utils/expectation-counters/countLines';
 import { countPages } from '../utils/expectation-counters/countPages';
 import { countParagraphs } from '../utils/expectation-counters/countParagraphs';
-import { splitIntoSentences } from '../utils/expectation-counters/countSentences';
-import { countSentences } from '../utils/expectation-counters/countSentences';
+import { countSentences, splitIntoSentences } from '../utils/expectation-counters/countSentences';
 import { countWords } from '../utils/expectation-counters/countWords';
 import { CountUtils } from '../utils/expectation-counters/index';
 import { extractParameterNames } from '../utils/extractParameterNames';
 import { capitalize } from '../utils/normalization/capitalize';
 import { decapitalize } from '../utils/normalization/decapitalize';
 import { DIACRITIC_VARIANTS_LETTERS } from '../utils/normalization/DIACRITIC_VARIANTS_LETTERS';
-import type { string_keyword } from '../utils/normalization/IKeywords';
-import type { IKeywords } from '../utils/normalization/IKeywords';
+import type { IKeywords, string_keyword } from '../utils/normalization/IKeywords';
 import { isValidKeyword } from '../utils/normalization/isValidKeyword';
 import { nameToUriPart } from '../utils/normalization/nameToUriPart';
 import { nameToUriParts } from '../utils/normalization/nameToUriParts';
@@ -54,6 +48,9 @@ import { $randomSeed } from '../utils/random/$randomSeed';
 import { removeEmojis } from '../utils/removeEmojis';
 import { removeQuotes } from '../utils/removeQuotes';
 import { replaceParameters } from '../utils/replaceParameters';
+import { clonePipeline } from '../utils/serialization/clonePipeline';
+import { deepClone } from '../utils/serialization/deepClone';
+import { $deepFreeze } from '../utils/serialization/deepFreeze';
 import { difference } from '../utils/sets/difference';
 import { intersection } from '../utils/sets/intersection';
 import { union } from '../utils/sets/union';
@@ -70,78 +67,81 @@ import { isUrlOnPrivateNetwork } from '../utils/validators/url/isUrlOnPrivateNet
 import { isValidPipelineUrl } from '../utils/validators/url/isValidPipelineUrl';
 import { isValidUrl } from '../utils/validators/url/isValidUrl';
 import { isValidUuid } from '../utils/validators/uuid/isValidUuid';
-
+import { PROMPTBOOK_VERSION } from '../version';
 
 // Note: Exporting version from each package
 export { PROMPTBOOK_VERSION };
 
-
 // Note: Entities of the `@promptbook/utils`
-export { renderPromptbookMermaid };
-export { extractParameterNamesFromPromptTemplate };
-export { extractVariables };
-export { renameParameter };
-export { titleToName };
-export { forEachAsync };
-export { isValidJsonString };
-export { extractBlock };
-export { $currentDate };
-export { clonePipeline };
-export { deepClone };
-export { $deepFreeze };
-export { $getGlobalScope };
-export { $isRunningInBrowser };
-export { $isRunningInNode };
-export { $isRunningInWebWorker };
-export { countCharacters };
-export { countLines };
-export { countPages };
-export { countParagraphs };
-export { splitIntoSentences };
-export { countSentences };
-export { countWords };
-export { CountUtils };
-export { extractParameterNames };
-export { capitalize };
-export { decapitalize };
-export { DIACRITIC_VARIANTS_LETTERS };
-export type { string_keyword };
-export type { IKeywords };
-export { isValidKeyword };
-export { nameToUriPart };
-export { nameToUriParts };
-export type { string_kebab_case };
-export { normalizeToKebabCase };
-export type { string_camelCase };
-export { normalizeTo_camelCase };
-export type { string_PascalCase };
-export { normalizeTo_PascalCase };
-export type { string_SCREAMING_CASE };
-export { normalizeTo_SCREAMING_CASE };
-export { normalizeTo_snake_case };
-export { normalizeWhitespaces };
-export { parseKeywords };
-export { parseKeywordsFromString };
-export { removeDiacritics };
-export { searchKeywords };
-export { parseNumber };
-export { $randomSeed };
-export { removeEmojis };
-export { removeQuotes };
-export { replaceParameters };
-export { difference };
-export { intersection };
-export { union };
-export { trimCodeBlock };
-export { trimEndOfCodeBlock };
-export { unwrapResult };
-export { isValidEmail };
-export { isValidFilePath };
-export { isValidJavascriptName };
-export { isValidPromptbookVersion };
-export { isValidSemanticVersion };
-export { isHostnameOnPrivateNetwork };
-export { isUrlOnPrivateNetwork };
-export { isValidPipelineUrl };
-export { isValidUrl };
-export { isValidUuid };
+export {
+    $currentDate,
+    $deepFreeze,
+    $getGlobalScope,
+    $isRunningInBrowser,
+    $isRunningInNode,
+    $isRunningInWebWorker,
+    $randomSeed,
+    capitalize,
+    clonePipeline,
+    countCharacters,
+    countLines,
+    countPages,
+    countParagraphs,
+    countSentences,
+    CountUtils,
+    countWords,
+    decapitalize,
+    deepClone,
+    DIACRITIC_VARIANTS_LETTERS,
+    difference,
+    extractBlock,
+    extractParameterNames,
+    extractParameterNamesFromPromptTemplate,
+    extractVariables,
+    forEachAsync,
+    intersection,
+    isHostnameOnPrivateNetwork,
+    isUrlOnPrivateNetwork,
+    isValidEmail,
+    isValidFilePath,
+    isValidJavascriptName,
+    isValidJsonString,
+    isValidKeyword,
+    isValidPipelineUrl,
+    isValidPromptbookVersion,
+    isValidSemanticVersion,
+    isValidUrl,
+    isValidUuid,
+    nameToUriPart,
+    nameToUriParts,
+    normalizeTo_camelCase,
+    normalizeTo_PascalCase,
+    normalizeTo_SCREAMING_CASE,
+    normalizeTo_snake_case,
+    normalizeToKebabCase,
+    normalizeWhitespaces,
+    parseKeywords,
+    parseKeywordsFromString,
+    parseNumber,
+    removeDiacritics,
+    removeEmojis,
+    removeQuotes,
+    renameParameter,
+    renderPromptbookMermaid,
+    replaceParameters,
+    searchKeywords,
+    splitIntoSentences,
+    titleToName,
+    trimCodeBlock,
+    trimEndOfCodeBlock,
+    union,
+    unwrapResult,
+};
+export type {
+    IKeywords,
+    string_camelCase,
+    string_kebab_case,
+    string_keyword,
+    string_PascalCase,
+    string_SCREAMING_CASE,
+};