+++ src/utils/serialization/deepFreeze.ts
@@ -1,5 +1,5 @@
 import type { ReadonlyDeep } from 'type-fest';
-import type { really_any } from './organization/really_any';
+import type { really_any } from '../organization/really_any';
 
 /**
  * @@@